Update for PHPUnit 9.x
