+++ composer.json
@@ -33,5 +33,8 @@
         "psr-4": {
             "Guzzle\\Service\\Loader\\": "src/"
         }
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^9"
     }
 }
